+++ src/test/java/com/atlassian/braid/BraidTest.java
@@ -1,0 +1,39 @@
+package com.atlassian.braid;
+
+import org.junit.Rule;
+import org.junit.Test;
+
+@SuppressWarnings("unchecked")
+public class BraidTest {
+
+    @Rule
+    public YamlBraidExecutionRule rule = new YamlBraidExecutionRule();
+
+    @Test
+    public void testBraidWithExtraVariables() {
+    }
+
+    @Test
+    public void testBraidWithTwoSchemaSources() {
+    }
+
+    @Test
+    public void testBraidWithSchemaSourceError() {
+    }
+
+    @Test
+    public void testBraidWithInlineVariables() {
+    }
+
+    @Test
+    public void testBraidWithFragment() {
+    }
+
+    @Test
+    public void testBraidWithFragmentAndVariables() {
+    }
+
+    @Test
+    public void testBraidWithMultipleNodesUsingFragment() {
+    }
+}