+++ CHANGES.md
@@ -4,7 +4,10 @@ Changelog for graphql-braid
 (Unreleased)
 -------------------
 
-- 
+- Support multiple use of fragments
+- Support multiple variables in different areas of a query
+- New config builder
+- Clean up functional tests
 
 0.3.1 (2017-11-20)
 -------------------